requirements.txt
digi-xbee>=1.1
wheel>=0.32
pyserial==3.4
